english-coach-api/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
python-multipart
